0054187 added additional fields + some refactorring
