+++ Controllers/Backend/viaebConfigForm.php
@@ -5,6 +5,7 @@ use Shopware\Components\CacheManager;
 use Shopware\Components\ConfigWriter;
 use viaebShopwareAfterbuy\Services\Helper\ShopwareConfigHelper;
 
+/** @noinspection PhpUnused */
 class Shopware_Controllers_Backend_viaebConfigForm extends Shopware_Controllers_Backend_ExtJs
 {
     /** @var ConfigWriter $configWriter */
@@ -46,7 +47,6 @@ class Shopware_Controllers_Backend_viaebConfigForm extends Shopware_Controllers_
         ]);
 
         try{
-            //TODO: typecast related to Shopware Version ordernumberMapping mainSystem ExportAllArticles
             $this->configWriter->save('partnerId', $_REQUEST['partnerId'], $this->pluginName);
             $this->configWriter->save('partnerPassword', $_REQUEST['partnerPassword'], $this->pluginName);
             $this->configWriter->save('userName', $_REQUEST['userName'], $this->pluginName);
