+++ Controllers/Backend/viaebResetShopConnection.php
@@ -3,6 +3,7 @@
 
 use viaebShopwareAfterbuy\Services\Helper\ShopwareResetHelper;
 
+/** @noinspection PhpUnused */
 class Shopware_Controllers_Backend_viaebResetShopConnection extends Shopware_Controllers_Backend_ExtJs
 {
     public function indexAction()
