+++ Services/Helper/ShopwareOrderHelper.php
@@ -4,9 +4,11 @@
 namespace viaebShopwareAfterbuy\Services\Helper;
 
 use DateTime;
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use Exception;
+use Shopware\Components\Model\ModelEntity;
 use Shopware\Models\Article\Detail as ArticleDetail;
 use Shopware\Models\Attribute\Order as OrderAttributes;
 use viaebShopwareAfterbuy\ValueObjects\Address as ValueAddress;
@@ -30,7 +32,6 @@ use Shopware\Models\Payment\Payment;
 
 class ShopwareOrderHelper extends AbstractHelper
 {
-
     protected $paymentStates;
 
     protected $shippingStates;
@@ -49,11 +50,17 @@ class ShopwareOrderHelper extends AbstractHelper
     /** @var ShopwareArticleHelper */
     protected $articleHelper;
 
+    /**
+     * @param ShopwareArticleHelper $articleHelper
+     */
     public function init(ShopwareArticleHelper $articleHelper)
     {
         $this->articleHelper = $articleHelper;
     }
 
+    /**
+     *
+     */
     public function preFetch()
     {
         $this->paymentStates = $this->getPaymentStates();
@@ -64,6 +71,9 @@ class ShopwareOrderHelper extends AbstractHelper
         $this->targetGroup = $this->getDefaultGroup();
     }
 
+    /**
+     * @return array
+     */
     public function getABCountryCodes()
     {
         return array(
@@ -382,6 +392,10 @@ class ShopwareOrderHelper extends AbstractHelper
 
     }
 
+    /**
+     * @param ShopwareOrder $order
+     * @return bool
+     */
     public function isFullfilled(ShopwareOrder $order)
     {
         $completelyPaid = $order->getPaymentStatus()->getId() === 12;
@@ -391,6 +405,9 @@ class ShopwareOrderHelper extends AbstractHelper
         return $completelyPaid && ($completelyDelivered || $completed);
     }
 
+    /**
+     * @return mixed
+     */
     public function getNewFullfilledOrders()
     {
         $lastExport = $this->entityManager->getRepository(Status::class)->find(1);
@@ -420,6 +437,9 @@ class ShopwareOrderHelper extends AbstractHelper
         return $orders;
     }
 
+    /**
+     * @return array
+     */
     public function getUnfullfilledOrders() {
 
         $orders = $this->entityManager->createQueryBuilder()
@@ -463,6 +483,10 @@ class ShopwareOrderHelper extends AbstractHelper
         return $this->shippingType;
     }
 
+    /**
+     * @param Order $value
+     * @param ShopwareOrder $order
+     */
     public function setPositions(ValueOrder $value, ShopwareOrder &$order)
     {
         $details = $order->getDetails();
@@ -514,6 +538,12 @@ class ShopwareOrderHelper extends AbstractHelper
         }
     }
 
+    /**
+     * @param Order $value
+     * @param ShopwareOrder $order
+     * @param Customer $customer
+     * @param string $type
+     */
     public function setAddress(ValueOrder $value, ShopwareOrder &$order, Customer $customer, $type = 'billing')
     {
         if ($type === 'billing') {
@@ -575,6 +605,10 @@ class ShopwareOrderHelper extends AbstractHelper
         }
     }
 
+    /**
+     * @param Order $value
+     * @param ShopwareOrder $order
+     */
     public function setOrderTaxValues(ValueOrder $value, ShopwareOrder &$order)
     {
         if ( ! $value->getAmountNet()) {
@@ -588,6 +622,11 @@ class ShopwareOrderHelper extends AbstractHelper
         }
     }
 
+    /**
+     * @param Order $value
+     * @param ShopwareOrder $order
+     * @param Shop $shop
+     */
     public function setOrderMainValues(ValueOrder $value, ShopwareOrder &$order, Shop $shop)
     {
         /**
@@ -634,6 +673,10 @@ class ShopwareOrderHelper extends AbstractHelper
         }
     }
 
+    /**
+     * @param Order $value
+     * @param ShopwareOrder $order
+     */
     public function setPaymentStatus(ValueOrder $value, ShopwareOrder &$order)
     {
         if ($value->getPaid() > 0) {
@@ -647,11 +690,18 @@ class ShopwareOrderHelper extends AbstractHelper
         }
     }
 
+    /**
+     * @param int $id
+     * @return object|null
+     */
     public function getShop(int $id)
     {
         return $this->entityManager->getRepository(Shop::class)->find($id);
     }
 
+    /**
+     * @return mixed
+     */
     public function getCountries()
     {
         $countries = $this->entityManager->createQueryBuilder()
@@ -663,6 +713,9 @@ class ShopwareOrderHelper extends AbstractHelper
         return $countries;
     }
 
+    /**
+     * @return mixed
+     */
     public function getPaymentStates()
     {
         $states = $this->entityManager->createQueryBuilder()
@@ -692,6 +745,9 @@ class ShopwareOrderHelper extends AbstractHelper
         return $states;
     }
 
+    /**
+     * @return mixed
+     */
     public function getDetailStates()
     {
         $states = $this->entityManager->createQueryBuilder()
@@ -717,6 +773,12 @@ class ShopwareOrderHelper extends AbstractHelper
         return $types;
     }
 
+    /**
+     * @param Order $order
+     * @param ValueAddress $billingAddress
+     * @param Shop $shop
+     * @return object|Customer|null
+     */
     public function getCustomer(ValueOrder $order, ValueAddress $billingAddress, Shop $shop)
     {
         $customer = $this->entityManager->getRepository(Customer::class)
@@ -729,6 +791,12 @@ class ShopwareOrderHelper extends AbstractHelper
         return $this->createCustomer($order, $billingAddress, $shop);
     }
 
+    /**
+     * @param Order $order
+     * @param ValueAddress $billingAddress
+     * @param Shop $shop
+     * @return Customer
+     */
     public function createCustomer(ValueOrder $order, ValueAddress $billingAddress, Shop $shop)
     {
         $customer = new Customer();
@@ -783,6 +851,9 @@ class ShopwareOrderHelper extends AbstractHelper
         return $customer;
     }
 
+    /**
+     * @return object|null
+     */
     public function getDefaultGroup()
     {
         return $this->entityManager->getRepository(Group::class)->findOneBy(array());
@@ -804,7 +875,7 @@ class ShopwareOrderHelper extends AbstractHelper
     }
 
     /**
-     * TODO: Optimize performance
+     *
      * @param Order[] $orders
      */
     public function resetArticleChangeTime(array $orders)
@@ -833,6 +904,111 @@ class ShopwareOrderHelper extends AbstractHelper
             $this->entityManager->flush();
         } catch (ORMException $e) {
             $this->logger->error($e->getMessage());
+        }
+    }
+
+    /**
+     * @param Billing|Shipping|ModelEntity $entity
+     * @return ValueAddress
+     */
+    public function buildAddress(ModelEntity $entity) {
+        $address = new ValueAddress();
+        $address->setFirstname($entity->getFirstName());
+        $address->setLastname($entity->getLastName());
+        $address->setCompany($entity->getCompany());
+        $address->setStreet($entity->getStreet());
+
+        if($entity->getAdditionalAddressLine1()) {
+            $address->setAdditionalAddressLine1($entity->getAdditionalAddressLine1());
+        }
+        $address->setZipcode($entity->getZipCode());
+        $address->setCity($entity->getCity());
+        $address->setCountry($entity->getCountry()->getIso());
+        $address->setPhone($entity->getPhone());
+
+        return $address;
+    }
+
+    /**
+     * @param ShopwareOrder|ModelEntity $entity
+     * @return ArrayCollection
+     */
+    public function buildPositions(ShopwareOrder $entity) {
+        $positions = new ArrayCollection();
+
+        foreach($entity->getDetails() as $position) {
+            /**
+             * @var Detail $position
+             */
+            $orderPosition = new OrderPosition();
+            if($position->getEan()) {
+                $orderPosition->setExternalIdentifier($position->getEan());
+            }
+
+            $orderPosition->setInternalIdentifier($position->getArticleNumber());
+            $orderPosition->setName($position->getArticleName());
+            $orderPosition->setPrice($position->getPrice());
+            $orderPosition->setQuantity($position->getQuantity());
+            $orderPosition->setTax($position->getTaxRate());
+
+            $positions->add($orderPosition);
+        }
+
+        return $positions;
+    }
+
+    /**
+     * @param Order $order
+     * @param ShopwareOrder $entity
+     */
+    public function setOrderValues(Order &$order, ShopwareOrder $entity) {
+        /** @noinspection PhpParamsInspection */
+        $order->setCreateDate($entity->getOrderTime());
+        $order->setShipping($entity->getInvoiceShipping());
+
+        try {
+            $order->setShippingType($entity->getDispatch()->getName());
+        }
+        catch(Exception $e) {
+            $order->setShippingType('Standard');
+        }
+
+        $order->setPaymentType($entity->getPayment()->getName());
+        $order->setPaymentTypeId($entity->getPayment()->getId());
+
+        if($entity->getTaxFree()) {
+            $order->setTaxFree(true);
+        }
+        $order->setInternalIdentifier($entity->getNumber());
+        $order->setCurrency($entity->getCurrency());
+        $order->setTransactionId($entity->getTransactionId());
+    }
+
+    /**
+     * @param Order $order
+     * @param ShopwareOrder $entity
+     */
+    public function setOrderStatus(Order &$order, ShopwareOrder $entity) {
+        if($entity->getPaymentStatus()->getId() === 12) {
+            $order->setPaid(true);
+            $order->setCleared(true);
+        }
+    }
+
+    /**
+     * @param Order $order
+     * @param ValueAddress $billingAddress
+     * @param ShopwareOrder $entity
+     */
+    public function setOrderCustomerData(Order &$order, ValueAddress &$billingAddress, ShopwareOrder $entity) {
+        if ($entity->getCustomer()) {
+            $billingAddress->setEmail($entity->getCustomer()->getEmail());
+            $order->setCustomerNumber($entity->getCustomer()->getNumber());
+        }
+
+        if ($entity->getCustomer() && $entity->getCustomer()->getBirthday()) {
+            /** @noinspection PhpParamsInspection */
+            $billingAddress->setBirthday($entity->getCustomer()->getBirthday());
         }
     }
 }
