+++ Services/ReadData/Internal/ReadOrdersService.php
@@ -4,13 +4,9 @@ namespace viaebShopwareAfterbuy\Services\ReadData\Internal;
 
 use Exception;
 use viaebShopwareAfterbuy\Services\Helper\ShopwareOrderHelper;
-use Doctrine\Common\Collections\ArrayCollection;
 use viaebShopwareAfterbuy\Services\ReadData\AbstractReadDataService;
 use viaebShopwareAfterbuy\Services\ReadData\ReadDataInterface;
-use viaebShopwareAfterbuy\ValueObjects\Address;
 use viaebShopwareAfterbuy\ValueObjects\Order;
-use viaebShopwareAfterbuy\ValueObjects\OrderPosition;
-use Shopware\Models\Order\Detail;
 use Shopware\Models\Order\Repository;
 
 class ReadOrdersService extends AbstractReadDataService implements ReadDataInterface {
@@ -43,110 +39,31 @@ class ReadOrdersService extends AbstractReadDataService implements ReadDataInter
         $targetData = array();
 
         foreach($data as $entity) {
-            /**
-             * @var \Shopware\Models\Order\Order $entity
-             */
-
             /** ignore order if not valid */
             if($entity->getBilling() === null || $entity->getDetails() === null) {
                 continue;
             }
 
-            /**
-             * @var Order $order
-             */
+            /** @var Order $order */
             $order = new $this->targetEntity();
 
-            $positions = new ArrayCollection();
-
-            foreach($entity->getDetails() as $position) {
-                /**
-                 * @var Detail $position
-                 */
-                $orderPosition = new OrderPosition();
-                if($position->getEan()) {
-                    $orderPosition->setExternalIdentifier($position->getEan());
-                }
-
-                $orderPosition->setInternalIdentifier($position->getArticleNumber());
-                $orderPosition->setName($position->getArticleName());
-                $orderPosition->setPrice($position->getPrice());
-                $orderPosition->setQuantity($position->getQuantity());
-                $orderPosition->setTax($position->getTaxRate());
-
-                $positions->add($orderPosition);
-            }
-
+            //set order positions
+            $positions = $this->helper->buildPositions($entity);
             $order->setPositions($positions);
 
-            $billingAddress = new Address();
-            $billingAddress->setFirstname($entity->getBilling()->getFirstName());
-            $billingAddress->setLastname($entity->getBilling()->getLastName());
-            $billingAddress->setCompany($entity->getBilling()->getCompany());
-            $billingAddress->setStreet($entity->getBilling()->getStreet());
+            //set address related information
+            $billingAddress = $this->helper->buildAddress($entity->getBilling());
+            $shippingAddress = $this->helper->buildAddress($entity->getShipping());
 
-            if($entity->getBilling()->getAdditionalAddressLine1()) {
-                $billingAddress->setAdditionalAddressLine1($entity->getBilling()->getAdditionalAddressLine1());
-            }
-            $billingAddress->setZipcode($entity->getBilling()->getZipCode());
-            $billingAddress->setCity($entity->getBilling()->getCity());
-            $billingAddress->setCountry($entity->getBilling()->getCountry()->getIso());
-            $billingAddress->setPhone($entity->getBilling()->getPhone());
-
-            if ($entity->getCustomer()) {
-                $billingAddress->setEmail($entity->getCustomer()->getEmail());
-                $order->setCustomerNumber($entity->getCustomer()->getNumber());
-            }
-
-            if ($entity->getCustomer() && $entity->getCustomer()->getBirthday()) {
-                /** @noinspection PhpParamsInspection */
-                $billingAddress->setBirthday($entity->getCustomer()->getBirthday());
-            }
-
-            $shippingAddress = new Address();
-            $shippingAddress->setFirstname($entity->getShipping()->getFirstName());
-            $shippingAddress->setLastname($entity->getShipping()->getLastName());
-            $shippingAddress->setCompany($entity->getShipping()->getCompany());
-            $shippingAddress->setStreet($entity->getShipping()->getStreet());
-            if($entity->getShipping()->getAdditionalAddressLine1()) {
-                $shippingAddress->setAdditionalAddressLine1($entity->getShipping()->getAdditionalAddressLine1());
-            }
-            $shippingAddress->setZipcode($entity->getShipping()->getZipCode());
-            $shippingAddress->setCity($entity->getShipping()->getCity());
-            $shippingAddress->setCountry($entity->getShipping()->getCountry()->getIso());
+            //set customer related data
+            $this->helper->setOrderCustomerData($order, $billingAddress, $entity);
 
             $order->setBillingAddress($billingAddress);
             $order->setShippingAddress($shippingAddress);
 
-            /** @noinspection PhpParamsInspection */
-            $order->setCreateDate($entity->getOrderTime());
-            $order->setShipping($entity->getInvoiceShipping());
-
-            try {
-                $shippingType = $entity->getDispatch();
-                $order->setShippingType($shippingType->getName());
-            }
-            catch(Exception $e) {
-                $order->setShippingType('Standard');
-            }
-
-            $order->setPaymentType($entity->getPayment()->getName());
-            $order->setPaymentTypeId($entity->getPayment()->getId());
-
-            if($entity->getTaxFree()) {
-                $order->setTaxFree(true);
-            }
-
-            $order->setInternalIdentifier($entity->getNumber());
-
-            $order->setCurrency($entity->getCurrency());
-
-            if($entity->getPaymentStatus()->getId() === 12) {
-                $order->setPaid(true);
-                $order->setCleared(true);
-            }
-
-            $order->setTransactionId($entity->getTransactionId());
+            //set values
+            $this->helper->setOrderValues($order, $entity);
+            $this->helper->setOrderStatus($order, $entity);
 
             $targetData[] = $order;
         }
